EPMLABSBRN-348 group csv parsing classes by content <eom>
