+++ src/main/kotlin/com/epam/brn/model/Series.kt
@@ -1,7 +1,7 @@
 package com.epam.brn.model
 
 import com.epam.brn.dto.SeriesDto
-import com.epam.brn.upload.csv.record.SeriesGenericRecord
+import com.epam.brn.upload.csv.series.SeriesGenericRecord
 import javax.persistence.CascadeType
 import javax.persistence.Column
 import javax.persistence.Entity
