+++ src/main/kotlin/com/epam/brn/upload/csv/MappingIteratorProvider.kt
@@ -1,4 +1,4 @@
-package com.epam.brn.upload.csv.parser.iterator
+package com.epam.brn.upload.csv
 
 import com.fasterxml.jackson.databind.MappingIterator
 import java.io.InputStream
