+++ src/main/kotlin/com/epam/brn/upload/csv/RecordProcessor.kt
@@ -1,4 +1,4 @@
-package com.epam.brn.upload.csv.processor
+package com.epam.brn.upload.csv
 
 interface RecordProcessor<Record, Entity> {
