+++ src/main/kotlin/com/epam/brn/upload/csv/group/GroupRecordMappingIteratorProvider.kt
@@ -1,7 +1,6 @@
-package com.epam.brn.upload.csv.parser.iterator.impl
+package com.epam.brn.upload.csv.group
 
-import com.epam.brn.upload.csv.parser.iterator.MappingIteratorProvider
-import com.epam.brn.upload.csv.record.SeriesGenericRecord
+import com.epam.brn.upload.csv.MappingIteratorProvider
 import com.fasterxml.jackson.databind.MappingIterator
 import com.fasterxml.jackson.dataformat.csv.CsvMapper
 import com.fasterxml.jackson.dataformat.csv.CsvParser
@@ -10,27 +9,28 @@ import org.apache.commons.lang3.StringUtils
 import org.springframework.stereotype.Component
 
 @Component
-class SeriesGenericRecordMappingIteratorProvider : MappingIteratorProvider<SeriesGenericRecord> {
+class GroupRecordMappingIteratorProvider :
+    MappingIteratorProvider<GroupRecord> {
 
-    override fun iterator(inputStream: InputStream): MappingIterator<SeriesGenericRecord> {
+    override fun iterator(inputStream: InputStream): MappingIterator<GroupRecord> {
         val csvMapper = CsvMapper().apply {
             enable(CsvParser.Feature.TRIM_SPACES)
         }
 
         val csvSchema = csvMapper
-            .schemaFor(SeriesGenericRecord::class.java)
+            .schemaFor(GroupRecord::class.java)
             .withColumnSeparator(',')
             .withLineSeparator(StringUtils.SPACE)
             .withColumnReordering(true)
             .withHeader()
 
         return csvMapper
-            .readerWithTypedSchemaFor(SeriesGenericRecord::class.java)
+            .readerWithTypedSchemaFor(GroupRecord::class.java)
             .with(csvSchema)
             .readValues(inputStream)
     }
 
     override fun isApplicable(format: String): Boolean {
-        return SeriesGenericRecord.FORMAT == format
+        return GroupRecord.FORMAT == format
     }
 }
