+++ src/main/kotlin/com/epam/brn/upload/csv/group/GroupRecordProcessor.kt
@@ -1,8 +1,8 @@
-package com.epam.brn.upload.csv.processor
+package com.epam.brn.upload.csv.group
 
 import com.epam.brn.model.ExerciseGroup
 import com.epam.brn.repo.ExerciseGroupRepository
-import com.epam.brn.upload.csv.record.GroupRecord
+import com.epam.brn.upload.csv.RecordProcessor
 import org.springframework.stereotype.Component
 import org.springframework.transaction.annotation.Transactional
