+++ src/main/kotlin/com/epam/brn/upload/csv/series/SeriesGenericRecordMappingIteratorProvider.kt
@@ -1,7 +1,6 @@
-package com.epam.brn.upload.csv.parser.iterator.impl
+package com.epam.brn.upload.csv.series
 
-import com.epam.brn.upload.csv.parser.iterator.MappingIteratorProvider
-import com.epam.brn.upload.csv.record.GroupRecord
+import com.epam.brn.upload.csv.MappingIteratorProvider
 import com.fasterxml.jackson.databind.MappingIterator
 import com.fasterxml.jackson.dataformat.csv.CsvMapper
 import com.fasterxml.jackson.dataformat.csv.CsvParser
@@ -10,27 +9,28 @@ import org.apache.commons.lang3.StringUtils
 import org.springframework.stereotype.Component
 
 @Component
-class GroupRecordMappingIteratorProvider : MappingIteratorProvider<GroupRecord> {
+class SeriesGenericRecordMappingIteratorProvider :
+    MappingIteratorProvider<SeriesGenericRecord> {
 
-    override fun iterator(inputStream: InputStream): MappingIterator<GroupRecord> {
+    override fun iterator(inputStream: InputStream): MappingIterator<SeriesGenericRecord> {
         val csvMapper = CsvMapper().apply {
             enable(CsvParser.Feature.TRIM_SPACES)
         }
 
         val csvSchema = csvMapper
-            .schemaFor(GroupRecord::class.java)
+            .schemaFor(SeriesGenericRecord::class.java)
             .withColumnSeparator(',')
             .withLineSeparator(StringUtils.SPACE)
             .withColumnReordering(true)
             .withHeader()
 
         return csvMapper
-            .readerWithTypedSchemaFor(GroupRecord::class.java)
+            .readerWithTypedSchemaFor(SeriesGenericRecord::class.java)
             .with(csvSchema)
             .readValues(inputStream)
     }
 
     override fun isApplicable(format: String): Boolean {
-        return GroupRecord.FORMAT == format
+        return SeriesGenericRecord.FORMAT == format
     }
 }
