+++ src/main/kotlin/com/epam/brn/upload/csv/series2/SeriesTwoRecordProcessor.kt
@@ -1,4 +1,4 @@
-package com.epam.brn.upload.csv.processor
+package com.epam.brn.upload.csv.series2
 
 import com.epam.brn.constant.ExerciseType
 import com.epam.brn.constant.WordType
@@ -9,7 +9,7 @@ import com.epam.brn.model.Task
 import com.epam.brn.repo.ExerciseRepository
 import com.epam.brn.repo.ResourceRepository
 import com.epam.brn.repo.SeriesRepository
-import com.epam.brn.upload.csv.record.SeriesTwoRecord
+import com.epam.brn.upload.csv.RecordProcessor
 import org.apache.commons.lang3.StringUtils
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Component
